Cleanup options, refactor transforms
