+++ v2/firestore-to-firestore/src/test/java/com/google/cloud/teleport/v2/templates/FirestoreToFirestoreTest.java
@@ -15,20 +15,12 @@
  */
 package com.google.cloud.teleport.v2.templates;
 
-import com.google.common.collect.ImmutableList;
-import com.google.firestore.v1.Document;
 import com.google.firestore.v1.DocumentRootName;
-import com.google.firestore.v1.MapValue;
 import com.google.firestore.v1.PartitionQueryRequest;
-import com.google.firestore.v1.RunQueryResponse;
 import com.google.firestore.v1.StructuredQuery;
 import com.google.firestore.v1.StructuredQuery.CollectionSelector;
-import com.google.firestore.v1.Value;
-import com.google.firestore.v1.Write;
-import java.util.Map;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.ValueProvider.StaticValueProvider;
-import org.apache.beam.sdk.testing.PAssert;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
@@ -42,113 +34,8 @@ import org.junit.runners.JUnit4;
 @RunWith(JUnit4.class)
 public class FirestoreToFirestoreTest {
 
-  @Rule public final transient TestPipeline p = TestPipeline.create();
-
-  @Test
-  public void testRunQueryResponseToDocument_extractsDocument() {
-    Document testDoc =
-        Document.newBuilder()
-            .setName("projects/test-project/databases/(default)/documents/myCol/doc1")
-            .build();
-    RunQueryResponse response = RunQueryResponse.newBuilder().setDocument(testDoc).build();
-
-    PCollection<RunQueryResponse> input = p.apply(Create.of(response));
-    PCollection<Document> output =
-        input.apply(ParDo.of(new FirestoreToFirestore.RunQueryResponseToDocument()));
-
-    PAssert.that(output).containsInAnyOrder(testDoc);
-    p.run();
-  }
-
-  @Test
-  public void testRunQueryResponseToDocument_skipsMissingDocument() {
-    RunQueryResponse response =
-        RunQueryResponse.newBuilder()
-            .setReadTime(com.google.protobuf.Timestamp.newBuilder().setSeconds(12345))
-            .build(); // No document
-
-    PCollection<RunQueryResponse> input = p.apply(Create.of(response));
-    PCollection<Document> output =
-        input.apply(ParDo.of(new FirestoreToFirestore.RunQueryResponseToDocument()));
-
-    PAssert.that(output).empty();
-    p.run();
-  }
-
-  @Test
-  public void testPrepareWritesFn_correctlyTransformsName() {
-    String sourceName = "projects/source-proj/databases/(default)/documents/myCol/docId1";
-    Document inputDoc =
-        Document.newBuilder()
-            .setName(sourceName)
-            .putFields("message", Value.newBuilder().setStringValue("Hello").build())
-            .build();
-
-    String destProject = "dest-proj";
-    String destDb = "dest-db";
-
-    PCollection<Document> input = p.apply(Create.of(inputDoc));
-    PCollection<Write> output =
-        input.apply(ParDo.of(new FirestoreToFirestore.PrepareWritesFn(destProject, destDb)));
-
-    String expectedName = "projects/dest-proj/databases/dest-db/documents/myCol/docId1";
-    Document expectedDoc = inputDoc.toBuilder().setName(expectedName).build();
-    Write expectedWrite = Write.newBuilder().setUpdate(expectedDoc).build();
-
-    PAssert.that(output).containsInAnyOrder(expectedWrite);
-    p.run();
-  }
-
-  @Test
-  public void testPrepareWritesFn_preservesFields() {
-    String sourceName = "projects/source/databases/(default)/documents/data/item1";
-    MapValue nestedMap =
-        MapValue.newBuilder()
-            .putFields("nested", Value.newBuilder().setBooleanValue(true).build())
-            .build();
-    Document inputDoc =
-        Document.newBuilder()
-            .setName(sourceName)
-            .putFields("fieldA", Value.newBuilder().setStringValue("valueA").build())
-            .putFields("fieldB", Value.newBuilder().setIntegerValue(123).build())
-            .putFields("fieldC", Value.newBuilder().setMapValue(nestedMap).build())
-            .build();
-
-    PCollection<Document> input = p.apply(Create.of(inputDoc));
-    PCollection<Write> output =
-        input.apply(ParDo.of(new FirestoreToFirestore.PrepareWritesFn("dest", "(default)")));
-
-    // Extract the Document from the Write
-    PCollection<Document> outputDocs =
-        output.apply("ExtractUpdate", ParDo.of(new ExtractDocumentFromWriteFn()));
-
-    // Assert that the fields map is as expected.
-    PCollection<Map<String, Value>> outputFields =
-        outputDocs.apply("ExtractFields", ParDo.of(new ExtractFieldsFn()));
-
-    // Pass the expected map as a List to PAssert
-    PAssert.that(outputFields).containsInAnyOrder(ImmutableList.of(inputDoc.getFieldsMap()));
-
-    p.run();
-  }
-
-  private static class ExtractDocumentFromWriteFn extends DoFn<Write, Document> {
-    @ProcessElement
-    public void processElement(ProcessContext c) {
-      if (c.element() != null && c.element().hasUpdate()) {
-        c.output(c.element().getUpdate());
-      }
-    }
-  }
-
-  private static class ExtractFieldsFn extends DoFn<Document, Map<String, Value>> {
-    @ProcessElement
-    public void processElement(ProcessContext c) {
-      if (c.element() != null) {
-        c.output(c.element().getFieldsMap());
-      }
-    }
-  }
+  @Rule
+  public final transient TestPipeline p = TestPipeline.create();
 
   // Note: Fully testing the pipeline logic with FirestoreIO requires integration tests
   // or a sophisticated mocking of FirestoreIO's builders and transforms, which is complex.
@@ -160,9 +47,9 @@ public class FirestoreToFirestoreTest {
 
     options.setFirestoreReadProjectId(StaticValueProvider.of("test-source-project"));
     options.setFirestoreReadDatabaseId(StaticValueProvider.of("test-source-db"));
+    options.setFirestoreReadCollection(StaticValueProvider.of("my-collection"));
     options.setFirestoreWriteProjectId(StaticValueProvider.of("test-dest-project"));
     options.setFirestoreWriteDatabaseId(StaticValueProvider.of("(default)"));
-    options.setFirestoreWriteEntityKind(StaticValueProvider.of("my-collection"));
 
     // Simply creating the pipeline and calling main to ensure it constructs without exceptions
     // We are not running it as it would try to connect to actual Firestore.
