+++ v2/firestore-to-firestore/src/test/java/com/google/cloud/teleport/v2/transforms/PrepareWritesFnTest.java
@@ -1,0 +1,103 @@
+package com.google.cloud.teleport.v2.transforms;
+
+import com.google.common.collect.ImmutableList;
+import com.google.firestore.v1.Document;
+import com.google.firestore.v1.MapValue;
+import com.google.firestore.v1.Value;
+import com.google.firestore.v1.Write;
+import java.util.Map;
+import org.apache.beam.sdk.testing.PAssert;
+import org.apache.beam.sdk.testing.TestPipeline;
+import org.apache.beam.sdk.transforms.Create;
+import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.values.PCollection;
+import org.junit.Rule;
+import org.junit.Test;
+
+public class PrepareWritesFnTest {
+
+  @Rule
+  public final transient TestPipeline p = TestPipeline.create();
+
+  @Test
+  public void testPrepareWritesFn_correctlyTransformsName() {
+    String sourceName = "projects/source-proj/databases/(default)/documents/myCol/docId1";
+    Document inputDoc =
+        Document.newBuilder()
+            .setName(sourceName)
+            .putFields("message", Value.newBuilder().setStringValue("Hello").build())
+            .build();
+
+    String destProject = "dest-proj";
+    String destDb = "dest-db";
+
+    PCollection<Document> input = p.apply(Create.of(inputDoc));
+    PCollection<Write> output =
+        input.apply(
+            ParDo.of(new PrepareWritesFn(destProject, destDb)));
+
+    String expectedName = "projects/dest-proj/databases/dest-db/documents/myCol/docId1";
+    Document expectedDoc = inputDoc.toBuilder().setName(expectedName).build();
+    Write expectedWrite = Write.newBuilder().setUpdate(expectedDoc).build();
+
+    PAssert.that(output).containsInAnyOrder(expectedWrite);
+    p.run();
+  }
+
+  @Test
+  public void testPrepareWritesFn_preservesFields() {
+    String sourceName = "projects/source/databases/(default)/documents/data/item1";
+    MapValue nestedMap =
+        MapValue.newBuilder()
+            .putFields("nested", Value.newBuilder().setBooleanValue(true).build())
+            .build();
+    Document inputDoc =
+        Document.newBuilder()
+            .setName(sourceName)
+            .putFields("fieldA", Value.newBuilder().setStringValue("valueA").build())
+            .putFields("fieldB", Value.newBuilder().setIntegerValue(123).build())
+            .putFields("fieldC", Value.newBuilder().setMapValue(nestedMap).build())
+            .build();
+
+    PCollection<Document> input = p.apply(Create.of(inputDoc));
+    PCollection<Write> output =
+        input.apply(
+            ParDo.of(new PrepareWritesFn("dest", "(default)")));
+
+    // Extract the Document from the Write
+    PCollection<Document> outputDocs =
+        output.apply("ExtractUpdate", ParDo.of(new ExtractDocumentFromWriteFn()));
+
+    // Assert that the fields map is as expected.
+    PCollection<Map<String, Value>> outputFields =
+        outputDocs.apply("ExtractFields", ParDo.of(new ExtractFieldsFn()));
+
+    // Pass the expected map as a List to PAssert
+    PAssert.that(outputFields).containsInAnyOrder(ImmutableList.of(inputDoc.getFieldsMap()));
+
+    p.run();
+  }
+
+  private static class ExtractDocumentFromWriteFn extends DoFn<Write, Document> {
+
+    @ProcessElement
+    public void processElement(ProcessContext c) {
+      if (c.element() != null && c.element().hasUpdate()) {
+        c.output(c.element().getUpdate());
+      }
+    }
+  }
+
+  private static class ExtractFieldsFn extends DoFn<Document, Map<String, Value>> {
+
+    @ProcessElement
+    public void processElement(ProcessContext c) {
+      if (c.element() != null) {
+        c.output(c.element().getFieldsMap());
+      }
+    }
+  }
+
+
+}
